Revert "refactor: auto retry for agent jobs"
