+++ press/patches.txt
@@ -106,4 +106,3 @@ press.press.doctype.virtual_machine.patches.rename_aws_fields
 press.press.doctype.virtual_machine_image.patches.rename_aws_fields
 press.press.doctype.virtual_disk_snapshot.patches.rename_aws_fields
 press.press.doctype.virtual_machine_volume.patches.rename_aws_fields
-press.press.doctype.agent_job.patches.update_status_for_undelivered_jobs
