+++ press/press/doctype/agent_job/agent_job.py
@@ -26,10 +26,6 @@ from press.press.doctype.press_notification.press_notification import (
 	create_new_notification,
 )
 
-from press.press.doctype.agent_job_type.agent_job_type import (
-	get_retryable_job_types_and_max_retry_count,
-)
-
 
 class AgentJob(Document):
 	dashboard_fields = [
@@ -114,16 +110,20 @@ class AgentJob(Document):
 			self.status = "Pending"
 			self.save()
 		except Exception:
-			if 400 <= cint(self.flags.status_code) <= 499:
-				self.status = "Failure"
-				self.save()
-				process_job_updates(self.name)
+			self.status = "Failure"
+			self.save()
 
-			else:
-				self.set_status_and_next_retry_at()
+			process_job_updates(self.name)
+
+			self.reload()
+			self.set_status_and_next_retry_at()
 
 	def set_status_and_next_retry_at(self):
+		if 400 <= cint(self.flags.status_code) <= 499:
+			return
+
 		try:
+
 			next_retry_at = get_next_retry_at(self.retry_count)
 
 			if not self.retry_count:
@@ -139,6 +139,7 @@ class AgentJob(Document):
 				},
 				update_modified=False,
 			)
+
 		except Exception:
 			log_error("Agent Job Set Status Exception", job=self)
 
@@ -548,14 +549,6 @@ def get_next_retry_at(job_retry_count):
 
 
 def retry_undelivered_jobs():
-	"""Retry undelivered jobs and update job status if max retry count is reached"""
-
-	disable_auto_retry = frappe.db.get_single_value(
-		"Press Settings", "disable_auto_retry", cache=True
-	)
-	if disable_auto_retry:
-		return
-
 	job_types, max_retry_per_job_type = get_retryable_job_types_and_max_retry_count()
 	server_jobs = get_server_wise_undelivered_jobs(job_types)
 
@@ -569,7 +562,10 @@ def retry_undelivered_jobs():
 
 		for job in undelivered_jobs:
 			job_doc = frappe.get_doc("Agent Job", job)
-			max_retry_count = max_retry_per_job_type[job_doc.job_type] or 0
+			max_retry_count = max_retry_per_job_type[job_doc.job_type]
+
+			if not max_retry_count:
+				continue
 
 			if job_doc.retry_count < max_retry_count:
 				retry = job_doc.retry_count + 1
@@ -577,6 +573,19 @@ def retry_undelivered_jobs():
 				job_doc.retry_in_place()
 			else:
 				update_job_and_step_status(job)
+
+
+def get_retryable_job_types_and_max_retry_count():
+	job_types, max_retry_per_job_type = [], {}
+	for job_type in frappe.get_all(
+		"Agent Job Type",
+		filters={"disabled_auto_retry": 0, "max_retry_count": [">", 0]},
+		fields=["name", "max_retry_count"],
+	):
+		job_types.append(job_type["name"])
+		max_retry_per_job_type[job_type["name"]] = job_type["max_retry_count"]
+
+	return job_types, max_retry_per_job_type
 
 
 def update_job_and_step_status(job):
