+++ press/press/doctype/site_migration/test_site_migration.py
@@ -10,17 +10,14 @@ from press.press.doctype.remote_file.remote_file import RemoteFile
 from frappe.tests.utils import FrappeTestCase
 from press.press.doctype.agent_job.agent_job import poll_pending_jobs
 from press.press.doctype.agent_job.test_agent_job import fake_agent_job
-from press.press.doctype.site.site import Site
+
 from press.press.doctype.site.test_site import create_test_bench, create_test_site
 
 
 @patch.object(RemoteFile, "download_link", new="http://test.com")
 class TestSiteMigration(FrappeTestCase):
 	def test_in_cluster_site_migration_goes_through_all_steps_and_updates_site(self):
-		with patch.object(Site, "after_insert"), patch.object(Site, "on_update"):
-			"""Patching these methods as its creating issue with duplicate agent job check"""
-			site = create_test_site()
-
+		site = create_test_site()
 		bench = create_test_bench()
 		site_migration = frappe.get_doc(
 			{
@@ -30,7 +27,7 @@ class TestSiteMigration(FrappeTestCase):
 			}
 		).insert()
 
-		with fake_agent_job("Update Site Configuration", "Success"), fake_agent_job(
+		with fake_agent_job("Update Site Configuration"), fake_agent_job(
 			"Backup Site",
 			data={
 				"backups": {
@@ -94,9 +91,7 @@ class TestSiteMigration(FrappeTestCase):
 		self.assertEqual(site.server, bench.server)
 
 	def test_site_is_activated_on_failure_when_possible(self):
-		with patch.object(Site, "after_insert"), patch.object(Site, "on_update"):
-			"""Patching these methods as its creating issue with duplicate agent job check"""
-			site = create_test_site()
+		site = create_test_site()
 		bench = create_test_bench()
 		site_migration = frappe.get_doc(
 			{
